Drag and drop now is implemented and in its own file.  The next step is to also have the user be able to click and open a file explorer to select a file that way.  Ideally, drag and drop and the button can be in 1 element, though the implementation is being a bit hard.
